region_segmentation/requirements.txt
absl-py==0.15.0
albumentations==1.1.0
asciitree==0.3.3
attrs==21.4.0
boto3==1.22.11
botocore==1.25.11
cachetools==4.2.4
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi==1.15.0
charset-normalizer==2.0.7
click==8.1.3
cloudpickle==2.0.0
ctk-cli==1.5
cycler==0.11.0
dask==2022.5.0
descartes==1.1.0
diskcache==5.4.0
distributed==2022.5.0
dynamic-watershed==1.0.1
efficientnet-pytorch==0.6.3
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
fasteners @ file:///tmp/build/80754af9/fasteners_1623953468841/work
fsspec @ file:///croot/fsspec_1679418786824/work
future==0.18.2
GDAL==3.5.0
girder-client==3.1.14
girder-slicer-cli-web==1.2.10
Glymur==0.9.9
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.41.1
HeapDict==1.0.1
histomicstk==1.2.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.31.1
imagesize @ file:///tmp/build/80754af9/imagesize_1637939814114/work
imgaug==0.4.0
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.5.1
kiwisolver==1.3.2
large-image==1.14.3
large-image-source-bioformats==1.14.3
large-image-source-deepzoom==1.14.3
large-image-source-dummy==1.14.3
large-image-source-gdal==1.14.3
large-image-source-mapnik==1.14.3
large-image-source-multi==1.14.3
large-image-source-nd2==1.14.3
large-image-source-ometiff==1.14.3
large-image-source-openjpeg==1.14.3
large-image-source-openslide==1.14.3
large-image-source-pil==1.14.3
large-image-source-test==1.14.3
large-image-source-tiff==1.14.3
large-image-source-vips==1.14.3
lazy_loader==0.2
locket==1.0.0
lxml==4.6.4
mapnik==4.0.0
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
msgpack @ file:///opt/conda/conda-bld/msgpack-python_1652362659880/work
munch==2.5.0
nd2==0.2.2
networkx==3.1
nimfa==1.4.0
numcodecs @ file:///croot/numcodecs_1674851371595/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1675642561072/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
openslide-python==1.1.2
packaging==21.3
palettable==3.3.0
pandas==1.4.2
partd==1.2.0
Pillow==9.5.0
pkgconfig==1.5.5
pretrainedmodels==0.7.4
prettytable==2.4.0
progressbar2==3.55.0
protobuf==3.19.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylibtiff==0.5.0
pyparsing==3.0.8
pyproj==3.3.1.1
pyrsistent==0.18.1
python-bioformats==4.0.5
python-dateutil==2.8.2
python-javabridge==4.0.3
python-utils==2.5.6
pytorch-warmup==0.0.4
pytz==2022.1
pyvips==2.2.0
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resource-backed-dask-array==0.1.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
scikit-image==0.21.0
scikit-learn==0.24.2
scipy==1.10.1
segmentation-models-pytorch==0.2.1
Shapely==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
tblib==1.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
threadpoolctl==3.0.0
tifffile==2023.4.12
tiffslide @ file:///home/conda/feedstock_root/build_artifacts/tiffslide_1657921537067/work
tifftools==1.3.4
timm==0.4.12
toolz==0.11.2
torch==1.10.1+cu113
torchaudio==0.10.1+cu113
torchfile==0.1.0
torchvision==0.11.2+cu113
tornado==6.1
tqdm==4.32.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.2
wurlitzer==3.0.2
xmltodict==0.12.0
zarr @ file:///croot/zarr_1669363532162/work
zict==2.2.0
